% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OSM.R
\name{OSMGetOutputRequestTypes}
\alias{OSMGetOutputRequestTypes}
\title{Provide the Different Outputs of OSM}
\usage{
OSMGetOutputRequestTypes()
}
\value{
a vector with the different outputs
}
\description{
Provide the Different Outputs of OSM
}
