% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OSM.R
\name{OSMGetVariantFields}
\alias{OSMGetVariantFields}
\title{Provide a List of Fields Required by a Variant}
\usage{
OSMGetVariantFields(variant)
}
\arguments{
\item{variant}{a valid variant (string)}
}
\value{
a list of data.frame instances containing the information on
the fields
}
\description{
Provide a List of Fields Required by a Variant
}
