% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OSM.R
\name{OSMGetVariantSpecies}
\alias{OSMGetVariantSpecies}
\title{Provide the species for}
\usage{
OSMGetVariantSpecies(variant, type = "All", outputAsVector = T)
}
\arguments{
\item{variant}{a valid variant (string)}

\item{type}{either All, Broadleaved, or Coniferous}

\item{outputAsVector}{a logical true provides a vector of species code}
}
\value{
a vector if outputAsVector is set to true or a list otherwise
}
\description{
Provide the species for
}
