% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OSM.R
\name{OSMSimulate}
\alias{OSMSimulate}
\title{Run a Simulation with OSM}
\usage{
OSMSimulate(data, outputRequestList, variant, years, ypc)
}
\arguments{
\item{data}{a data.frame instance properly formatted}

\item{outputRequestList}{an OutputRequestList instance}

\item{variant}{a valid variant (string)}

\item{years}{an integer that stands for the simulation length (yrs)}

\item{ypc}{the step length (yrs)}
}
\value{
a SimulationResult instance
}
\description{
Run a Simulation with OSM
}
