% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OSM.R
\name{OSMStatus}
\alias{OSMStatus}
\title{Provide the Status of OSM Web API}
\usage{
OSMStatus()
}
\value{
a list of data.frame instances containing the information on
the fields
}
\description{
Provide the Status of OSM Web API
}
